fix for https://issues.apache.org/activemq/browse/AMQ-910
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@812941 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-ra/src/main/java/org/apache/activemq/ra/ActiveMQActivationSpec.java
@@ -587,7 +587,7 @@ public class ActiveMQActivationSpec implements MessageActivationSpec, Serializab
         }
     }
 
-    public short getBackOffMultiplier() {
+    public double getBackOffMultiplier() {
         if (redeliveryPolicy == null) {
             return 0;
         }
